Client: Was renamed subpackage from "config" to "core". Server: Was added detail comments for JavaDoc.
