+++ client/src/main/java/ru/shishmakov/client/core/HttpClientProcessorHandler.java
@@ -1,4 +1,4 @@
-package ru.shishmakov.client.config;
+package ru.shishmakov.client.core;
 
 
 import io.netty.channel.ChannelHandler.Sharable;
