+++ client/src/main/java/ru/shishmakov/client/core/PackageMarker.java
@@ -1,4 +1,4 @@
-package ru.shishmakov.client.config;
+package ru.shishmakov.client.core;
 
 /**
  * Empty marker class.<br/>
